requirements.txt
playwright>=1.41.0
pydantic>=2.5.3
pydantic-ai>=0.0.19
python-dotenv>=1.0.0
anthropic>=0.8.1 